// serverjs/src/protocol/sidecar-codec.d.ts
export declare function encodeSidecar<T>(schemaId: number, value: T): Uint8Array;
export declare function decodeSidecar<T>(schemaId: number, bytes: Uint8Array): T | null;
export declare const SIDECAR_SCHEMA: Readonly<{
    INT: 1;
    STRING: 2;
    JWT_LOGIN: 3;
    JWT_RESPONSE: 4;
    ITEM: 10;
    BULK_ITEMS: 11;
    DELETE_ITEM: 12;
    BULK_DELETE_ITEMS: 13;
    MOVE_ITEM: 14;
    CHARACTER_SELECT: 20;
    PLAYER_PROFILE: 21;
    CHARACTER_CREATE: 22;
    ENTER_WORLD: 23;
    ZONE_SESSION: 30;
    ZONE_CHANGE: 31;
    NEW_ZONE: 32;
    SPAWN: 40;
    SPAWNS: 41;
    ENTITY_POSITION: 42;
    CLIENT_POSITION: 43;
    ANIMATION: 44;
    LEVEL: 45;
    CHANNEL: 46;
    COMMAND: 47;
    DELETE_SPAWN: 48;
}>;